Move all codey code into folder pkg
