+++ orion/pkg/router/handler.go
@@ -3,7 +3,7 @@ package router
 import (
     "github.com/gin-gonic/contrib/static"
     "github.com/gin-gonic/gin"
-    "github.com/ahsu1230/mathnavigatorSite/orion/controllers"
+    "github.com/ahsu1230/mathnavigatorSite/orion/pkg/controllers"
 )
 
 type Handler struct {
